--- a/tools.go
+++ b/tools.go
@@ -1,7 +1,0 @@
-//+build tools
-
-package flux
-
-import (
-	_ "github.com/mna/pigeon"
-)