There are a few notable differences.

First, true and false are not reserved identifiers so they will evaluate
as identifiers. It is up to the caller to set the value for these
predeclared identifiers. This is done manually in one test and otherwise
it is done automatically in the interpreter.

Second, some of the location information changed. Notably, object
expressions start at the `{` symbol instead of the location of the first
property and call expression start at the beginning of the callee
instead of the first parenthesis. This second one is identical to what
the go parser does. The type inference tests had to be modified for this
reason.

The parser options have been removed. They seem to have been
automatically generated by pigeon and weren't applicable anymore. This
changes the type signature, but it doesn't change how it is called in a
majority of situations.

The pigeon dependency has been removed as it is no longer needed.