@jsternberg 2018-12-04
Here's the benchmark results.

```
benchmark                                                                         old ns/op     new ns/op     delta
BenchmarkParser/optional_query_metadata-12                                        2565885       241426        -90.59%
BenchmarkParser/optional_query_metadata_preceding_query_text-12                   1815026       225428        -87.58%
BenchmarkParser/from-12                                                           324054        74126         -77.13%
BenchmarkParser/comment-12                                                        342353        74595         -78.21%
BenchmarkParser/identifier_with_number-12                                         337864        73827         -78.15%
BenchmarkParser/regex_literal-12                                                  248830        61212         -75.40%
BenchmarkParser/regex_literal_with_escape_sequence-12                             246759        60393         -75.53%
BenchmarkParser/regex_match_operators-12                                          1117767       136872        -87.75%
BenchmarkParser/declare_variable_as_an_int-12                                     620951        78324         -87.39%
BenchmarkParser/declare_variable_as_a_float-12                                    716647        84936         -88.15%
BenchmarkParser/declare_variable_as_an_array-12                                   3366540       114773        -96.59%
BenchmarkParser/declare_variable_as_an_empty_array-12                             748598        99404         -86.72%
BenchmarkParser/use_variable_to_declare_something-12                              875735        149501        -82.93%
BenchmarkParser/variable_is_from_statement-12                                     1096152       187297        -82.91%
BenchmarkParser/pipe_expression-12                                                612792        117086        -80.89%
BenchmarkParser/literal_pipe_expression-12                                        505982        119839        -76.32%
BenchmarkParser/member_expression_pipe_expression-12                              754446        132086        -82.49%
BenchmarkParser/multiple_pipe_expressions-12                                      1384018       163326        -88.20%
BenchmarkParser/two_variables_for_two_froms-12                                    1875691       261610        -86.05%
BenchmarkParser/from_with_database-12                                             926102        117676        -87.29%
BenchmarkParser/map_member_expressions-12                                         2006370       216951        -89.19%
BenchmarkParser/index_expression-12                                               626884        74858         -88.06%
BenchmarkParser/nested_index_expression-12                                        1044731       106579        -89.80%
BenchmarkParser/access_indexed_object_returned_from_function_call-12              624382        88484         -85.83%
BenchmarkParser/index_with_member_expressions-12                                  648429        97241         -85.00%
BenchmarkParser/index_with_member_with_call_expression-12                         796893        109039        -86.32%
BenchmarkParser/var_as_binary_expression_of_other_vars-12                         1681376       215497        -87.18%
BenchmarkParser/var_as_unary_expression_of_other_vars-12                          875333        129048        -85.26%
BenchmarkParser/var_as_both_binary_and_unary_expressions-12                       1187785       154116        -87.02%
BenchmarkParser/unary_expressions_within_logical_expression-12                    2191019       197051        -91.01%
BenchmarkParser/unary_expressions_with_too_many_comments-12                       2151241       211926        -90.15%
BenchmarkParser/expressions_with_function_calls-12                                822261        112886        -86.27%
BenchmarkParser/mix_unary_logical_and_binary_expressions-12                       1730294       149033        -91.39%
BenchmarkParser/mix_unary_logical_and_binary_expressions_with_extra_parens-12     2157465       157564        -92.70%
BenchmarkParser/arrow_function_called-12                                          2080402       214447        -89.69%
BenchmarkParser/arrow_function_return_map-12                                      2535541       141052        -94.44%
BenchmarkParser/arrow_function_with_default_arg-12                                2138884       155095        -92.75%
BenchmarkParser/arrow_function_called_in_binary_expression-12                     5210329       229653        -95.59%
BenchmarkParser/arrow_function_as_single_expression-12                            2281622       144057        -93.69%
BenchmarkParser/arrow_function_as_block-12                                        3295785       205324        -93.77%
BenchmarkParser/from_with_filter_with_no_parens-12                                7147968       374324        -94.76%
BenchmarkParser/from_with_range-12                                                2484638       209743        -91.56%
BenchmarkParser/from_with_limit-12                                                2046603       202324        -90.11%
BenchmarkParser/from_with_range_and_count-12                                      3476733       251152        -92.78%
BenchmarkParser/from_with_range,_limit_and_count-12                               4916671       309883        -93.70%
BenchmarkParser/from_with_join-12                                                 10097132      654997        -93.51%
BenchmarkParser/from_with_join_with_complex_expression-12                         18710547      975802        -94.78%
BenchmarkParser/duration_literal,_all_units-12                                    1244645       98099         -92.12%
BenchmarkParser/duration_literal,_months-12                                       462103        86247         -81.34%
BenchmarkParser/duration_literal,_milliseconds-12                                 510626        88215         -82.72%
BenchmarkParser/duration_literal,_months,_minutes,_milliseconds-12                647247        89644         -86.15%

benchmark                                                                         old allocs     new allocs     delta
BenchmarkParser/optional_query_metadata-12                                        5167           1054           -79.60%
BenchmarkParser/optional_query_metadata_preceding_query_text-12                   3439           892            -74.06%
BenchmarkParser/from-12                                                           787            241            -69.38%
BenchmarkParser/comment-12                                                        817            243            -70.26%
BenchmarkParser/identifier_with_number-12                                         787            241            -69.38%
BenchmarkParser/regex_literal-12                                                  647            207            -68.01%
BenchmarkParser/regex_literal_with_escape_sequence-12                             757            217            -71.33%
BenchmarkParser/regex_match_operators-12                                          1418           550            -61.21%
BenchmarkParser/declare_variable_as_an_int-12                                     884            287            -67.53%
BenchmarkParser/declare_variable_as_a_float-12                                    887            288            -67.53%
BenchmarkParser/declare_variable_as_an_array-12                                   3748           514            -86.29%
BenchmarkParser/declare_variable_as_an_empty_array-12                             1061           287            -72.95%
BenchmarkParser/use_variable_to_declare_something-12                              1403           439            -68.71%
BenchmarkParser/variable_is_from_statement-12                                     1459           578            -60.38%
BenchmarkParser/pipe_expression-12                                                774            379            -51.03%
BenchmarkParser/literal_pipe_expression-12                                        814            337            -58.60%
BenchmarkParser/member_expression_pipe_expression-12                              903            429            -52.49%
BenchmarkParser/multiple_pipe_expressions-12                                      1099           662            -39.76%
BenchmarkParser/two_variables_for_two_froms-12                                    2493           1058           -57.56%
BenchmarkParser/from_with_database-12                                             1894           444            -76.56%
BenchmarkParser/map_member_expressions-12                                         4264           983            -76.95%
BenchmarkParser/index_expression-12                                               2474           292            -88.20%
BenchmarkParser/nested_index_expression-12                                        4150           395            -90.48%
BenchmarkParser/access_indexed_object_returned_from_function_call-12              1640           343            -79.09%
BenchmarkParser/index_with_member_expressions-12                                  1185           388            -67.26%
BenchmarkParser/index_with_member_with_call_expression-12                         1002           439            -56.19%
BenchmarkParser/var_as_binary_expression_of_other_vars-12                         3285           965            -70.62%
BenchmarkParser/var_as_unary_expression_of_other_vars-12                          1514           534            -64.73%
BenchmarkParser/var_as_both_binary_and_unary_expressions-12                       2056           634            -69.16%
BenchmarkParser/unary_expressions_within_logical_expression-12                    3438           895            -73.97%
BenchmarkParser/unary_expressions_with_too_many_comments-12                       3596           900            -74.97%
BenchmarkParser/expressions_with_function_calls-12                                1425           441            -69.05%
BenchmarkParser/mix_unary_logical_and_binary_expressions-12                       3803           654            -82.80%
BenchmarkParser/mix_unary_logical_and_binary_expressions_with_extra_parens-12     7602           664            -91.27%
BenchmarkParser/arrow_function_called-12                                          4395           884            -79.89%
BenchmarkParser/arrow_function_return_map-12                                      6703           599            -91.06%
BenchmarkParser/arrow_function_with_default_arg-12                                3377           685            -79.72%
BenchmarkParser/arrow_function_called_in_binary_expression-12                     5510           1128           -79.53%
BenchmarkParser/arrow_function_as_single_expression-12                            3651           655            -82.06%
BenchmarkParser/arrow_function_as_block-12                                        6823           942            -86.19%
BenchmarkParser/from_with_filter_with_no_parens-12                                16144          1750           -89.16%
BenchmarkParser/from_with_range-12                                                2286           988            -56.78%
BenchmarkParser/from_with_limit-12                                                2733           928            -66.04%
BenchmarkParser/from_with_range_and_count-12                                      2313           1179           -49.03%
BenchmarkParser/from_with_range,_limit_and_count-12                               3167           1503           -52.54%
BenchmarkParser/from_with_join-12                                                 16112          3072           -80.93%
BenchmarkParser/from_with_join_with_complex_expression-12                         34736          4620           -86.70%
BenchmarkParser/duration_literal,_all_units-12                                    1747           345            -80.25%
BenchmarkParser/duration_literal,_months-12                                       821            296            -63.95%
BenchmarkParser/duration_literal,_milliseconds-12                                 804            296            -63.18%
BenchmarkParser/duration_literal,_months,_minutes,_milliseconds-12                1070           308            -71.21%

benchmark                                                                         old bytes     new bytes     delta
BenchmarkParser/optional_query_metadata-12                                        191096        32193         -83.15%
BenchmarkParser/optional_query_metadata_preceding_query_text-12                   147184        27500         -81.32%
BenchmarkParser/from-12                                                           38705         8214          -78.78%
BenchmarkParser/comment-12                                                        39704         8263          -79.19%
BenchmarkParser/identifier_with_number-12                                         38711         8214          -78.78%
BenchmarkParser/regex_literal-12                                                  35748         7206          -79.84%
BenchmarkParser/regex_literal_with_escape_sequence-12                             40117         7766          -80.64%
BenchmarkParser/regex_match_operators-12                                          60910         18054         -70.36%
BenchmarkParser/declare_variable_as_an_int-12                                     49539         9503          -80.82%
BenchmarkParser/declare_variable_as_a_float-12                                    44238         9503          -78.52%
BenchmarkParser/declare_variable_as_an_array-12                                   160728        16076         -90.00%
BenchmarkParser/declare_variable_as_an_empty_array-12                             44579         9519          -78.65%
BenchmarkParser/use_variable_to_declare_something-12                              62394         14043         -77.49%
BenchmarkParser/variable_is_from_statement-12                                     62725         18150         -71.06%
BenchmarkParser/pipe_expression-12                                                38992         12289         -68.48%
BenchmarkParser/literal_pipe_expression-12                                        41042         10985         -73.23%
BenchmarkParser/member_expression_pipe_expression-12                              42947         13755         -67.97%
BenchmarkParser/multiple_pipe_expressions-12                                      51453         20624         -59.92%
BenchmarkParser/two_variables_for_two_froms-12                                    98089         32345         -67.02%
BenchmarkParser/from_with_database-12                                             70195         14164         -79.82%
BenchmarkParser/map_member_expressions-12                                         163237        30039         -81.60%
BenchmarkParser/index_expression-12                                               111151        9695          -91.28%
BenchmarkParser/nested_index_expression-12                                        178415        12698         -92.88%
BenchmarkParser/access_indexed_object_returned_from_function_call-12              72983         11209         -84.64%
BenchmarkParser/index_with_member_expressions-12                                  57644         12529         -78.26%
BenchmarkParser/index_with_member_with_call_expression-12                         46844         14051         -70.00%
BenchmarkParser/var_as_binary_expression_of_other_vars-12                         140612        29591         -78.96%
BenchmarkParser/var_as_unary_expression_of_other_vars-12                          72933         16821         -76.94%
BenchmarkParser/var_as_both_binary_and_unary_expressions-12                       92015         19704         -78.59%
BenchmarkParser/unary_expressions_within_logical_expression-12                    136794        27086         -80.20%
BenchmarkParser/unary_expressions_with_too_many_comments-12                       141254        27246         -80.71%
BenchmarkParser/expressions_with_function_calls-12                                66140         13955         -78.90%
BenchmarkParser/mix_unary_logical_and_binary_expressions-12                       146106        20113         -86.23%
BenchmarkParser/mix_unary_logical_and_binary_expressions_with_extra_parens-12     277692        20353         -92.67%
BenchmarkParser/arrow_function_called-12                                          172401        27061         -84.30%
BenchmarkParser/arrow_function_return_map-12                                      237521        18591         -92.17%
BenchmarkParser/arrow_function_with_default_arg-12                                136258        21152         -84.48%
BenchmarkParser/arrow_function_called_in_binary_expression-12                     216564        34095         -84.26%
BenchmarkParser/arrow_function_as_single_expression-12                            126766        20161         -84.10%
BenchmarkParser/arrow_function_as_block-12                                        253705        29691         -88.30%
BenchmarkParser/from_with_filter_with_no_parens-12                                493255        52790         -89.30%
BenchmarkParser/from_with_range-12                                                89863         30241         -66.35%
BenchmarkParser/from_with_limit-12                                                105443        28303         -73.16%
BenchmarkParser/from_with_range_and_count-12                                      92800         35840         -61.38%
BenchmarkParser/from_with_range,_limit_and_count-12                               124371        46321         -62.76%
BenchmarkParser/from_with_join-12                                                 524540        92069         -82.45%
BenchmarkParser/from_with_join_with_complex_expression-12                         1079483       137152        -87.29%
BenchmarkParser/duration_literal,_all_units-12                                    83614         12785         -84.71%
BenchmarkParser/duration_literal,_months-12                                       45742         9831          -78.51%
BenchmarkParser/duration_literal,_milliseconds-12                                 39225         9831          -74.94%
BenchmarkParser/duration_literal,_months,_minutes,_milliseconds-12                50543         10480         -79.27%
```